Documentation + Kubernetes deployment scripts
This commit updates the Kubernetes deployment scripts to include new options for deploying, generating, transforming, importing, and querying datasets. The `deploy-converg.sh` script now accepts command line arguments (`--deploy`, `--generation`, `--transformation`, `--import`, `--query`) to specify which components and jobs to deploy. Similarly, the `delete-converg.sh` script now accepts the same command line arguments to specify which components and jobs to delete.

These changes improve the flexibility and ease of use of the deployment process for the Blazegraph and ConVer-G components.